SCC-32 Refactore code and decompose conf files.
