+++ CMakeLists.txt
@@ -8,7 +8,8 @@ add_compile_options(-Wall -Wextra)
 include_directories(include)
 add_executable(SCC
         src/main.cpp
-        src/conf.cpp
+        src/config.cpp
+        src/log.cpp
         src/lexical_analyzer.cpp
         src/syntax_analyzer.cpp
         src/ast_handler.cpp
