+++ src/config.cpp
@@ -1,8 +1,7 @@
-#include "SCC/conf.h"
+#include "SCC/config.h"
 
 using namespace std;
 
-Log SCC_log;
 Config config;
 
 void Configure(int argc, char* argv[]) {
@@ -18,73 +17,13 @@ void Configure(int argc, char* argv[]) {
   LOG(TRACE, "Configuration is completed");
 }
 
-string Log::get_timestamp() {
-  time_t now;
-  time(&now);
-  now += 3 * 3600;    // UTC+3 -- MSC
-  char buf[sizeof "2011-10-08T07:07:09"];
-  strftime(buf, sizeof buf, "%FT%T", gmtime(&now));
+string Config::GetConfigPath() const {
+  string cwf_path = __FILE__;
+  string cwf = cwf_path.substr(cwf_path.find_last_of('/') + 1);
+  string path = cwf_path.substr(0, cwf_path.find(cwf));
   
-  return buf;
+  return path;
 }
-string Log::TimeToLogName(string timestamp) {
-  replace(begin(timestamp), end(timestamp), ':', '-');
-  return timestamp + ".log";
-}
-void Log::ValidateLogLevel(LogLevel level) const {
-  try {
-    if (LogLevel::LOG_LEVEL_COUNT <= level) {
-      throw invalid_argument("incorrect SCC log level: " + to_string(level));
-    }
-  } catch (invalid_argument& ex) {
-    cout << ex.what();
-    exit(EXIT_FAILURE);
-  }
-}
-void Log::ValidateLogLevel(std::string& level) const {
-  try {
-    if (str2log_lvl_.count(level) == 0) {
-      throw invalid_argument("incorrect SCC log level: " + level);
-    }
-  } catch (invalid_argument& ex) {
-    LOG(ERROR, ex.what());
-    exit(EXIT_FAILURE);
-  }
-  LOG(DEBUG, "log level is valid");
-}
-
-Log::Log() {
-  filename_ = TimeToLogName(get_timestamp());
-  output_.open("../log/" + filename_, ios::out);
-}
-void Log::add_log(LogLevel level, const string& msg) {
-  ValidateLogLevel(level);
-  ostringstream output;
-  output << "[" << log_lvl2str_.at(level) << "]\t"
-         << get_timestamp() << " " << msg;
-  
-  output_ << output.str();
-  if (log_level_ >= level) {
-    cout << output.str();
-  }
-}
-LogLevel Log::get_log_level() const {
-  return log_level_;
-}
-void Log::set_log_level(LogLevel level) {
-  log_level_ = level;
-}
-LogLevel Log::StringToLogLevel(string& level) {
-  for_each(begin(level), end(level),
-           [](char& c) {
-    c = (char)::toupper(c);
-  });
-  ValidateLogLevel(level);
-  return str2log_lvl_.at(level);
-}
-Log::~Log() = default;
-
-
 bool Config::IsFileExists(string& path) {
   struct stat buffer{};
   return (stat(path.c_str(), &buffer) == 0);
@@ -93,7 +32,7 @@ bool Config::IsFileExists(const ofstream& f) {
   return f.good();
 }
 void Config::ValidateMode(SCCMode mode) const {
-  if (SCCMode::MODE_COUNT <= mode) {
+  if (SCCMode::SCCMODE_COUNT <= mode) {
     LOG(ERROR, "incorrect SCC mode: " << to_string(mode));
     exit(EXIT_FAILURE);
   }
@@ -106,25 +45,25 @@ void Config::ValidateMode(string& mode) const {
   }
   LOG(DEBUG, "mode is valid");
 }
-void Config::ValidateSQLPath(string& sql_path) const{
-  if (!IsFileExists(sql_path)) {
+void Config::ValidateSQLPath(string& sql_path) const {
+  if (!(this->IsFileExists(sql_path))) {
     LOG(ERROR, "SQL file does not exist: " << sql_path);
     exit(EXIT_FAILURE);
   }
   LOG(DEBUG, "SQL path is valid");
 }
 void Config::ValidateCypherPath(string& cypher_path) const {
-  if (!IsFileExists(output_)) {
+  if (!(this->IsFileExists(output_))) {
     LOG(ERROR, "Cypher file does not exist: " << cypher_path);
     exit(EXIT_FAILURE);
   }
   LOG(DEBUG, "Cypher path is valid");
 }
-bool Config::IsFlagSet(OptFlag flag) {
+bool Config::IsFlagSet(OptFlag flag) const {
   return is_config_set_.at(flag_to_config_.at(flag));
 }
-void Config::ValidateSetFlag(OptFlag flag) {
-  if (IsFlagSet(flag)) {
+void Config::ValidateSetFlag(OptFlag flag) const {
+  if (this->IsFlagSet(flag)) {
     LOG(ERROR, "conflicting flags");
     exit(EXIT_FAILURE);
   }
@@ -134,12 +73,8 @@ void Config::SetFlag(OptFlag flag) {
 }
 
 Config::Config() {
-  string cwf_path = __FILE__;
-  string cwf = cwf_path.substr(cwf_path.find_last_of('/') + 1);
-  string path = cwf_path.substr(0, cwf_path.find(cwf));
-
   mode_ = SCCMode::INTERACTIVE;
-  sql_path_ = cypher_path_ = path;
+  sql_path_ = cypher_path_ = this->GetConfigPath();
   sql_path_ += "../resources/sql_queries.sql";
   cypher_path_ += "../resources/cypher_queries.cypher";
 }
@@ -149,19 +84,17 @@ void Config::Start() {
   ValidateCypherPath(cypher_path_);
   // TODO: implement SCC mode behavior.
 }
-
 void Config::set_mode(SCCMode mode) {
-  ValidateMode(mode);
+  this->ValidateMode(mode);
   mode_ = mode;
 }
-void Config::set_sql_path(std::string& new_sql_path) {
-  ValidateSQLPath(new_sql_path);
+void Config::set_sql_path(string& new_sql_path) {
+  this->ValidateSQLPath(new_sql_path);
   sql_path_ = new_sql_path;
 }
-void Config::set_cypher_path(std::string& new_cypher_path) {
+void Config::set_cypher_path(string& new_cypher_path) {
   cypher_path_ = new_cypher_path;
 }
-
 SCCMode Config::get_mode() const {
   return mode_;
 }
@@ -171,23 +104,22 @@ string Config::get_sql_path() const {
 string Config::get_cypher_path() const {
   return cypher_path_;
 }
-ifstream& Config::Input() {
+ifstream& Config::ReadSQL() {
   return input_;
 }
-ofstream& Config::Output() {
+ofstream& Config::WriteCypher() {
   return output_;
 }
-
-SCCMode Config::StringToSCCMode(std::string& mode) {
+SCCMode Config::StringToSCCMode(string& mode) const {
   for_each(begin(mode), end(mode),
            [](char& c) {
     c = (char)::toupper(c);
   });
-  ValidateMode(mode);
+  this->ValidateMode(mode);
   return str2modes_.at(mode);
 }
-string Config::SCCModeToString(SCCMode mode) {
-  ValidateMode(mode);
+string Config::SCCModeToString(SCCMode mode) const {
+  this->ValidateMode(mode);
   return modes2str_.at(mode);
 }
 void Config::GetConsoleArguments(int argc, char* const* argv) {
@@ -215,53 +147,57 @@ void Config::GetConsoleArguments(int argc, char* const* argv) {
     auto OF_flag = static_cast<OptFlag>(flag);
     switch (flag) {
       case OptFlag::fDAEMON:
-        ValidateSetFlag(OF_flag);
-        LOG(TRACE, "set SCC mode = " << config.SCCModeToString(SCCMode::DAEMON));
-        config.set_mode(SCCMode::DAEMON);
-        SetFlag(OF_flag);
+        this->ValidateSetFlag(OF_flag);
+        LOG(TRACE, "set SCC mode = "
+                   << this->SCCModeToString(SCCMode::DAEMON));
+        this->set_mode(SCCMode::DAEMON);
+        this->SetFlag(OF_flag);
         break;
       case OptFlag::fHELP:
         LOG(INFO, "TODO: Get help");
         exit(EXIT_SUCCESS);
       case OptFlag::fINTERACTIVE:
-        ValidateSetFlag(OF_flag);
-        LOG(TRACE, "set SCC mode = " << config.SCCModeToString(SCCMode::INTERACTIVE));
-        config.set_mode(SCCMode::INTERACTIVE);
-        SetFlag(OF_flag);
+        this->ValidateSetFlag(OF_flag);
+        LOG(TRACE, "set SCC mode = "
+                   << this->SCCModeToString(SCCMode::INTERACTIVE));
+        this->set_mode(SCCMode::INTERACTIVE);
+        this->SetFlag(OF_flag);
         break;
       case OptFlag::fLOG:
+        this->ValidateSetFlag(OF_flag);
         LOG(TRACE, "set log level = " << optarg);
         {
           string tmp_log_level = optarg;
-          config.set_log_level(config.StringToLogLevel(tmp_log_level));
+          SCC_log.set_log_level(this->StringToLogLevel(tmp_log_level));
         }
+        this->SetFlag(OF_flag);
         break;
       case OptFlag::fMODE:
-        ValidateSetFlag(OF_flag);
+        this->ValidateSetFlag(OF_flag);
         LOG(TRACE, "set SCC mode = " << optarg);
         {
           string tmp_mode = optarg;
-          config.set_mode(config.StringToSCCMode(tmp_mode));
+          this->set_mode(this->StringToSCCMode(tmp_mode));
         }
-        SetFlag(OF_flag);
+        this->SetFlag(OF_flag);
         break;
       case OptFlag::fSQL:
-        ValidateSetFlag(OF_flag);
+        this->ValidateSetFlag(OF_flag);
         LOG(TRACE, "set SQL path = " << optarg);
         {
           string tmp = optarg;
-          config.set_sql_path(tmp);
+          this->set_sql_path(tmp);
         }
-        SetFlag(OF_flag);
+        this->SetFlag(OF_flag);
         break;
       case OptFlag::fCYPHER:
-        ValidateSetFlag(OF_flag);
+        this->ValidateSetFlag(OF_flag);
         LOG(TRACE, "set Cypher path = " << optarg);
         {
           string tmp = optarg;
-          config.set_cypher_path(tmp);
+          this->set_cypher_path(tmp);
         }
-        SetFlag(OF_flag);
+        this->SetFlag(OF_flag);
         break;
       default:
         char f = short_options[optind+1];
